RAYNCAMANA-402: Made chart preview scalable
